legal/AI/Ai_yusa/requirements.txt
# requirements.txt
# Ai_yusa - Gemini API 전용 경량 버전

# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3

# Data (판례 데이터 있을 때만 사용)
pandas>=2.1.0
pyarrow>=14.0.0

# LLM (Gemini)
google-generativeai>=0.3.2

# Utils
python-dotenv>=1.0.0
